requirements.txt
robotframework
selenium
webdriver-manager
robotframework-seleniumlibrary